refactor: IndexPage, TagTemplate 페이지 추상화 (#17)
